+++ tests/testthat.R
@@ -1,4 +1,4 @@
 library("testthat")
-
 library("arules")
+
 test_check("arules")