+++ tests/testthat/test-confint.R
@@ -1,9 +1,4 @@
-library("arules")
-library("testthat")
-
 debug <- FALSE
-
-context("confint")
 
 data("Adult")
 ## Mine association rules.