+++ tests/testthat/test-extract.R
@@ -1,12 +1,6 @@
-library("arules")
-library("testthat")
-
-context("extract")
-
 data("Adult")
 
 expect_equal(Adult, c(Adult[1:100], Adult[-(1:100)]))
-
 
 ec <- eclat(Adult,control=list(verb=FALSE))
 expect_equal(ec, c(ec[1:100], ec[-(1:100)]))