+++ tests/testthat/test-missing.R
@@ -1,5 +1,4 @@
 ## FIXME: missing tests
-## * itemCoding
 ## * aggregate
 ## * dissimilarity
 ## * pmml